Cast size_t to unsigned long to print with sprintf.

+++ src/tinyspline.c
@@ -162,7 +162,8 @@ tsError ts_internal_bspline_access_ctrlp_at(const tsBSpline *spline,
 	if (index >= ts_bspline_num_control_points(spline)) {
 		TS_RETURN_2(status, TS_INDEX_ERROR,
 			   "index (%lu) >= num(control_points) (%lu)",
-			   index, ts_bspline_num_control_points(spline))
+			    (unsigned long) index,
+			    (unsigned long) ts_bspline_num_control_points(spline))
 	}
 	*ctrlp = ts_internal_bspline_access_ctrlp(spline) +
 		index * ts_bspline_dimension(spline);
@@ -183,7 +184,7 @@ tsError ts_internal_bspline_generate_knots(const tsBSpline *spline,
 	if (type == TS_BEZIERS && n_knots % order != 0) {
 		TS_RETURN_2(status, TS_NUM_KNOTS,
 			   "num(knots) (%lu) %% order (%lu) != 0",
-			   n_knots, order)
+			    (unsigned long) n_knots, (unsigned long) order)
 	}
 
 	knots = ts_internal_bspline_access_knots(spline);
@@ -286,7 +287,8 @@ tsError ts_bspline_set_degree(tsBSpline *spline, size_t deg, tsStatus *status)
 	if (deg >= ts_bspline_num_control_points(spline)) {
 		TS_RETURN_2(status, TS_DEG_GE_NCTRLP,
 			   "degree (%lu) >= num(control_points) (%lu)",
-			   deg, ts_bspline_num_control_points(spline))
+			    (unsigned long) deg,
+			    (unsigned long) ts_bspline_num_control_points(spline))
 	}
 	spline->pImpl->deg = deg;
 	TS_RETURN_SUCCESS(status)
@@ -302,7 +304,8 @@ tsError ts_bspline_set_order(tsBSpline *spline, size_t order, tsStatus *status)
 	if (order == 0 || order > ts_bspline_num_control_points(spline)) {
 		TS_RETURN_2(status, TS_DEG_GE_NCTRLP,
 			   "order (%lu) > num(control_points) (%lu)",
-			   order, ts_bspline_num_control_points(spline))
+			    (unsigned long) order,
+			    (unsigned long) ts_bspline_num_control_points(spline))
 	}
 	return ts_bspline_set_degree(spline, order - 1, status);
 }
@@ -320,7 +323,8 @@ tsError ts_bspline_set_dimension(tsBSpline *spline, size_t dim,
 	if (ts_bspline_len_control_points(spline) % dim != 0) {
 		TS_RETURN_2(status, TS_LCTRLP_DIM_MISMATCH,
 			   "len(control_points) (%lu) %% dimension (%lu) != 0",
-			   ts_bspline_len_control_points(spline), dim)
+			    (unsigned long) ts_bspline_len_control_points(spline),
+			    (unsigned long) dim)
 	}
 	spline->pImpl->dim = dim;
 	TS_RETURN_SUCCESS(status)
@@ -430,14 +434,16 @@ tsError ts_bspline_set_knots(tsBSpline *spline, const tsReal *knots,
 			mult++;
 		} else if (lst_knot > knot) {
 			TS_RETURN_1(status, TS_KNOTS_DECR,
-				   "decreasing knot vector at index: %lu", idx)
+				   "decreasing knot vector at index: %lu",
+				    (unsigned long) idx)
 		} else {
 			mult = 0;
 		}
 		if (mult > order) {
 			TS_RETURN_3(status, TS_MULTIPLICITY,
 				   "mult(%f) (%lu) > order (%lu)",
-				   knot, mult, order)
+				   knot, (unsigned long) mult,
+				   (unsigned long) order)
 		}
 		lst_knot = knot;
 	}
@@ -571,13 +577,14 @@ tsError ts_bspline_new(size_t num_control_points, size_t dimension,
 	}
 	if (num_knots > TS_MAX_NUM_KNOTS) {
 		TS_RETURN_2(status, TS_NUM_KNOTS,
-			   "unsupported number of knots: %lu > %i", num_knots,
-			   TS_MAX_NUM_KNOTS)
+			   "unsupported number of knots: %lu > %i",
+			    (unsigned long) num_knots, TS_MAX_NUM_KNOTS)
 	}
 	if (degree >= num_control_points) {
 		TS_RETURN_2(status, TS_DEG_GE_NCTRLP,
 			   "degree (%lu) >= num(control_points) (%lu)",
-			   degree, num_control_points)
+			    (unsigned long) degree,
+			    (unsigned long) num_control_points)
 	}
 
 	_spline_->pImpl = (struct tsBSplineImpl *) malloc(sof_spline);
@@ -797,7 +804,7 @@ tsError ts_internal_relaxed_uniform_cubic_bspline(const tsReal *points,
 		TS_RETURN_0(status, TS_DIM_ZERO, "unsupported dimension: 0")
 	if (n <= 1) {
 		TS_RETURN_1(status, TS_DEG_GE_NCTRLP,
-			   "num(points) (%lu) <= 1", n)
+			   "num(points) (%lu) <= 1", (unsigned long) n)
 	}
 	/* in the following n >= 2 applies */
 
@@ -1500,7 +1507,7 @@ tsError ts_internal_bspline_from_json(const JSON_Value *spline_value,
 	if (len_ctrlp % dim != 0) {
 		TS_RETURN_2(status, TS_PARSE_ERROR,
 			   "len(control_points) (%lu) %% dimension (%lu) != 0",
-			   len_ctrlp, dim)
+			    (unsigned long) len_ctrlp, (unsigned long) dim)
 	}
 
 	/* Read number of knots. */
@@ -1519,7 +1526,8 @@ tsError ts_internal_bspline_from_json(const JSON_Value *spline_value,
 		if (num_knots != ts_bspline_num_knots(_spline_))
 			TS_THROW_2(try, status, TS_NUM_KNOTS,
 				   "unexpected num(knots): (%lu) != (%lu)",
-				   num_knots, ts_bspline_num_knots(_spline_))
+				   (unsigned long) num_knots,
+				   (unsigned long) ts_bspline_num_knots(_spline_))
 
 		/* Set control points. */
 		ctrlp = ts_internal_bspline_access_ctrlp(_spline_);
@@ -1528,7 +1536,7 @@ tsError ts_internal_bspline_from_json(const JSON_Value *spline_value,
 			if (json_value_get_type(real_value) != JSONNumber)
 				TS_THROW_1(try, status, TS_PARSE_ERROR,
 					   "control_points: value at index %lu is not a number",
-					   i)
+					   (unsigned long) i)
 			ctrlp[i] = (tsReal) json_value_get_number(real_value);
 		}
 		TS_CALL(try, err, ts_bspline_set_control_points(
@@ -1541,7 +1549,7 @@ tsError ts_internal_bspline_from_json(const JSON_Value *spline_value,
 			if (json_value_get_type(real_value) != JSONNumber)
 			TS_THROW_1(try, status, TS_PARSE_ERROR,
 				   "knots: value at index %lu is not a number",
-				   i)
+				   (unsigned long) i)
 			knots[i] = (tsReal) json_value_get_number(real_value);
 		}
 		TS_CALL(try, err, ts_bspline_set_knots(
